backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()

HOST = os.getenv("HOST", "0.0.0.0")
PORT = int(os.getenv("PORT", 8000))
OLLAMA_BASE_URL = os.getenv("OLLAMA_BASE_URL", "http://localhost:11434")

# Comma-separated allowed origins — use "*" only during local development
ALLOWED_ORIGINS: list[str] = os.getenv("ALLOWED_ORIGINS", "http://localhost:5173").split(",")

MAX_FILE_SIZE_MB = int(os.getenv("MAX_FILE_SIZE_MB", "50"))

MODEL_DEFAULTS = {
    "openai": "gpt-4o-mini",
    "gemini": "gemini-2.0-flash",
    "llama":  "llama3.2",
}
