pyproject.toml
[project]
name = "rag-chatbot"
version = "2.0.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "langchain==0.3.7",
    "langchain-openai==0.2.9",
    "langchain-google-genai==2.0.5",
    "langchain-ollama==0.2.1",
    "langchain-community==0.3.7",
    "langchain-text-splitters==0.3.4",
    "faiss-cpu==1.9.0",
    "chromadb==0.5.23",
    "onnxruntime<1.24",
    "pypdf==4.3.1",
    "beautifulsoup4==4.12.3",
    "lxml==5.3.0",
    "requests==2.32.3",
    "httpx==0.27.2",
    "python-multipart==0.0.12",
    "pydantic==2.9.2",
    "python-dotenv==1.0.1",
    "tiktoken==0.8.0",
    "sentence-transformers==3.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "ruff==0.8.6",
]

[tool.ruff]
line-length = 100
